Please believe me that it is not a pleasure or hobby of mine to alter the AR5 branch!
This time we change two things:

(1) in the radiation code, the RCOMPX call to diagnose OCII+OCIA+SOAs has
been split into two calls: OCII+OCIA and SOAs so that SOA rad forcing
and optical depths can be diagnosed separately. They maintain the same
optical properties, so only (small) change to climate forcing is by virtue
of the radiation code seeing the same tracer mass but separated into two
rad code tracers now. Altered interface to photolysis code to expect extra
tracer.

USERS of these diags must be AWARE of this change, to account for OCI, OCB,
and SOAs separately now when comparing runs before and after the recent
split of OCII+OCIA+OCB+SOAs into: OCII+OCIA, OCB, SOA diags.

(2) We are adding a sink of CO via O1D reaction, believing
this will alleviate a high CO bias in the stratosphere/upper troposphere.
So we are changing the climate simulation here with tracers on, but in
testing, any changes to tracer radiative forcing did not look worrisome.
Larissa is going to see what this does to the controls though.