The following changes were necessary when trying to run with FVCUBED=YES:
When using ESMF4.x there is a runtime error related to libesmf.so
that can be resolved by linking directly to the static library libesmf.a

ALso the number of GEOS5 linked libraries is incomplete causing linking errors.
Specifically MAPL_cfio, GEOS_Shared and GMAO_mpeu were added at the linking
state (note that the order was also important).