For NEW_IO_SUBDD:
Finally got around to changing the logic so that output
files are not opened/closed every timestep.  This required
the creation of a new subroutine close_subdd() which is called
at the end of execution of the main program (even for the
GISS-format option).

Also, I haven't tried to modify open_subdd(), firstly since I
don't understand everything it's doing, and secondly because
the metadata for output fields is not available in there.
Instead, the netcdf write routine simply checks whether the
output file for a particular quantity is open, and defines/opens
it if not.

If the main program crashes/aborts, the status of the netcdf subdd
files will be undetermined.  (This apparently was not an issue
for Fortran's I/O library.)

Please let me know of any problems.