Introduced a lean python script that will convert a fixed format file to free format.  It is fairly robust, but there are some obscure language constructs that can trip it up.   Such issues must be dealt with manually.
The original file provided a number of other modernization actions
including changing the case of keywords and automating the
indentation.  This new version _only_ deals with comment characters
and continuation lines and attempts to preserve indentation.  A second
pass with Emacs can be used to enforce modelE coding standards for
indentation.