Template coding for min/max diagnostics on the aij and oij domains.  The new acc-file arrays are called aijmm and oijmm, respectively.  aijmm is currently populated with the minimum/maximum instantaneous surface air temperatures, and oijmm with the maximum instantaneous KPP boundary layer depth.
The min/max operations can, of course, be applied to multi-hour
or other time averages instead of instantaneous values;
future work on the so-called "subdaily" diagnostics should
make it easier to automatically obtain these averages
by request.

It is not necessary to output both the minimum and maximum
for a field.

Acc-files are combined as one would expect; the min/max
are taken over files.

For convenience, minima and maxima are both stored in a
single accumulation array.  If this proves too confusing
(see below), I will change it. Minima are obtained using
the max() operator and the identity min(a,b) == -max(-a,-b).
The external minus sign is placed into the scale
factor, and the "accumulation" step for a minimum
diagnostic is written as
aijmm(i,j,ij_qtymin) = max(aijmm(i,j,ij_qtymin), -qty)
Future work on the diagnostics will make it possible
to hide these details.

The new arrays are only read/written/postprocessed for
the NEW_IO configuration; if necessary I can add the
coding for the traditional i/o.