More work to make the SEAICE_DRV.f routines flexible enough to handle the case that lake ice and ocean ice are on separate grids: the diagnostic accumulations in MELT_SI, PRECIP_SI, GROUND_SI, and FORM_SI have been collected into a single procedure SI_DIAGS. Once the lake and ocean ice grids differ, SI_DIAGS will manage the placement of the SI-scheme inputs/outputs into the applicable model E diagnostic arrays.
No changes to results other than roundoff differences in the diagnostics.

I thought it best to focus upcoming grid-abstraction changes in the
SI drivers themselves upon whatever is required to manage prognostic
variables only.

Placing the accumulations into a separate driver should not
significantly increase the risk that diagnostics are not
appropriately redefined when schemes are updated, since
most of the SI diagnostics are calculated from flux-coupler
arrays or are fundamental scheme-independent quantities
like temperature, salinity, and mass.  And the one-step
accumulation in SI_DIAGS, combined with some regrouping of
terms, makes it somewhat easier to see the definitions of
certain outputs and the correspondences between the
full-grid and regional/budget-page versions.