This is an attempt to simplify the "setup" process with the goal to make it more flexible and less confusing.
Attention!!!
Makefile functionality has changed !!!

Now "make setup ..." just compiles the model and creates the run
directory. It does *not* run the model for the first hour.
If you want the old functionality you have to use "make setup-run ..."
instead.

This was done to allow users to use "runpbs" script of their choice
to run the model (even for the first hour). The script should be able
to pass a flag "-cold-restart" to "E" script in run directory.
When such flag is passed the model will do initial setup run
(will use values from ISTART=... line in the rundeck). Without
this flag the model will do a normal restart from fort.[12] .

The script modelE/exec/runE was updated so that it can be used
to start a simple run: serial or mpi (if working on a node
that allows execution of mpirun). It also respects QSUB_STRING
set in the ~/.modelErc . If QSUB_STRING is set than runE
will submit a pbs job as specified by QSUB_STRING .
The command "make setup-run ...", actually, executes runE script.

The typical way to run runE script is

  runE <run_id> -np <number_of_cpus> -cold-restart

for initial start of the model. And

  runE <run_id> -np <number_of_cpus>

for normal restarts from fort.[12] .

When creating "I" file setup_e.pl now splits INPUTZ namelist
into two: INPUTZ and INPUTZ_cold. INPUTZ_cold contains only
the lines which start from "ISTART=..." and is read only
when flag -cold-restart is specified.