Added to AR5_branch set-up functionality similar to the one which was added recently to the master branch (i.e. "setup" and first hour run can be done separately now).
To prevent interference with current AR5 runs everything was
added as an option, i.e. separate files were committed
for the makefile and scripts. Among "standard" AR5 files only
MODELE.f was modified to add an extra option. To use this
new functionality you have to switch to the new makefile
by adding "-f new.mk" flag to your "make" commands. I.e.
typical way to set up the run and run it for the first hour
will be:

  make -f new.mk setup RUN=<RunID> ESMF=YES
  ../exec/runE_new <RunID> -cold-restart -np <number_of_cpus>

This functionality is mostly relevant to those interested
to run the model on Ames machines. See commit message
on master branch for more details.