1. Fixed reproducibility problem for subdaily diagnostics of turbulent deposition and gravitational settling of dust aerosols in some cases of restarts. The problem occured when a surface type was converted to another in the time step before the restart. The old value remained in the memory and was added again to the accumulation array of the subdaily variable at the next time step in the no-restart case. In case of a restart, no false value was added. Thus, the restart case was correct, the other one wasn't.
2. Fixed another bug in the accumulation of the subdaily dust
diagnostics. A variable was used outside the tracer loop, although it
was supposed to be used inside the tracer loop.

3. Fixed adiurn/hdiurn diagnostics also, which suffered from the same
mistake.

4. Problem: Uninitialized elements in the halo of variable datar8 in
get_subdd led to a crash of the model due to overflow, when the
content of this variable was assigned to variable data and transformed
to real(kind=4). This seemed to be happening using ifort 11. I don't
recall any problem with ifort 10. Solution: Initialization of datar8
and data to Zero at the start of the subroutine.

All these fixes affect only subdaily or adiurn/hdiurn diagnostics
specifically for dust aerosols. No change in the results, except for
some of the subdd or adiurn/hdiurn dust diagnostics.