The call to geopar() was moved to alloc_ocean().
This permits ocean geometry information to be initialized before it
is needed by standalone ocean configurations.

Due to the unfortunate combination of incomplete parallelization
of hycom and the lack of distinctions made between geometric
versus time-varying arrays, the move required a slight change
to the read-from-restart-file procedure to prevent gather_hycom_arrays
from clobbering the global-domain geometry information calculated in
geopar().