Expanding the code slightly to allow use of MERRA winds. To run, in addition to the NUDGE_ON CPP directive, add NUDGE_MERRA, and list the wind files in the same manner as we've done for NCEP.
The code changes are:
 (1) NUDGE_MERRA chooses 42 levels to read from the nc files instead
of 17 (NCEP), and
 (2) MERRA (unlike NCEP) has "missing data" values for pressures that
would be "below ground" so to speak. So there is a bit of code that
sets the lowest level for searching to the first that is defined,
rather than level 1 for NCEP.

pl8 dimensioned at nlevnc (the number of reanalysis levels) rather
than LM. Probably caused no problems for NCEP with model levels
 .gt. 17, but I noticed problems because nlevnc=42 for MERRA is
.gt. LM for 40 layer model.

In the process, made these improvements:
 (1) put in some error handling on the netCDF file interactions,
because I found one could run the model for weeks with bad winds
until the sea ice got too cold :-) and
 (2) corrected the closing of files.

On the AR5 code, MERRA nudging has been run multiple decades. Please
let me know if you have any problems on the master branch code.