First git-committed version of the QUS advection code for ocean R.
To activate, set ocean_use_qus=1 in the rundeck.

Adjustment of the moments in the vertical mixing scheme has
been implemented.  For flexibility and tidiness, this is now
performed _after_ the execution of the mixing code, using saved
diffusivites and fluxes. When using the default advection scheme,
the rewritten moment update gives the same result as before.
For the QUS case, there are two versions of the adjustment of
the vertical moments: (1) relaxation toward parabolic
subgrid profiles determined from triplets of layer-mean values, and
(2) application of the diffusion operator to a triple-resolution
version of the vertical profile created from the vertical moments,
followed by reconstruction of the vertical moments from the
diffused profile.  Method (1) has been used in all simulations
to date, but method (2) has been made the default and is
currently undergoing testing.

QUS-appropriate changes will soon be committed to OCNGM.f.