WIP commit for more flexibility in the administration of surface types.
- PBL arrays were moved to the derived type containing atm-surface coupling
arrays.  This change will permit multiple PBL instances per surface type when
needed.  A few of the PBL arrays _might_ be considered scheme-specific and
deserving their own data structures, but it is convenient to keep all the
arrays together for now.

- A number of arrays were placed into "bundles" for efficient
en-masse averaging over surface types.  The bundles are currently
implemented using some primitive but effective code added to
shared/ArrayBundle_mod.F90, to be upgraded later.

- Quantities averaged over surface types, e.g. the globally visible
array TSAVG from the FLUXES module, are now accessed as elements of
an instance of the derived type for atm-surface coupling
(at the moment, atmsrf in the FLUXES module contains the gridbox averages).