Several fixes needed for the model to work with the NAG 5.3 compiler. Most are for standard conformance - e.g. cannot pass scalar actual arguments to array dummy arguments when there is an explicit interface.
One fix in CLOUDS2.F90 is a kludge to workaround an internal compiler
error with NAG 5.3 on OS X.  One simple formula is split into two
statements.  No idea why this helps, but it does.

EC12 now runs 1 hr and 1dy on NAG 5.3.  We'll proceed to testing more
challenging rundecks.