Added -mismatch=all for F90FLAGS for NAG.
Simple solution to these unused unallocated allocatable
arrays.   By changing the initial "size" from 0 to -1, the first iteration
will trigger an allocate even when the computed size is just 0.

Fix for previous change that enabled portability for NAG.   The original fix
was for cases where unallocated allocatable arrays were passed to routines expecting
assumed size arrays.  The arrays are never actually used, so this is mostly a
matter of conforming to the standard.

Note that both of these fixes are not in the best location, but I could not
immediately understand how to extend the logic up in dd2d_utils.f where the
changes should go.  I will work on this, but am committing now so that the
model will be able to run.

MAX and MIN intrinsic require KIND of both arguments to be the same,
but some compilers allow it.  Fixed for NAG portability.

float() is not standard conforming. NAG requires real().

Eliminated per-variable SAVE attributes.  NAG complains that the
global SAVE conflicts with the per-variable SAVE's.

Added "EXTENDED_SOURCE" definition for NAG configuration.

Moved portability kludge for F2008 execute_command_line() to System.f.
System.f is the proper home for such system dependent items.

Fixing a logic error in my previous fix for NAG portability.
Also improved the clarity a tad since I was changing things anyway.

Minor fix for previous NAG fix.  0-sized array more correctly
expresses that dummy is being used only as a placeholder.

NAG port (and fixes from previous change for NAG).