WIP commit of modifications to support subdivisions of surface types and broader objectives related to atmosphere-surface interaction.
Main changes:

- atmosphere-surface interaction fields for a given surface
type are now stored in arrays dimensioned by the number of
subdivisions for that type.

- added an extra dimension to land ice arrays.  This currently
only serves as a way to test code correctness with multiple
atm.-landice flux calculations per gridbox.

- Eliminated the temporary use of pointers to arrays containing
gridbox-mean fields.  Such fields are now explicitly copied to
the per-surface-type data structures (downscaling will
replace copying when appropriate).  These operations are
performed in subroutines invoked at the appropriate stages
of execution.

- pulled diagnostic and landice-specific coding out of SURFACE.

- moved some snow and runoff variables into the coupling data structure

- pulled some generic diagnostics out of GHY_DRV, stored snow exports
in the coupling data structure

- FLOWO and related arrays now exist only in the atm-ocean coupling data
structure (as part of the reduction in the use of pointers to arrays
in the FLUXES module).

- most fields in the atm-surf coupling data structures are "bundled", i.e.
existing as part of a group of quantities playing a similar functional role
in the code (e.g. outputs of the PBL scheme).

- streamlined the setting of pointers in asflx arrays