First steps to make OCNGM.f more plug-and-play (to facilitate comparisons with the results of other numerical formulations of mesoscale processes).
1. Switched the proportionality coefficient so that thickness diffusivity
is made proportional to isoneutral diffusivity rather than the other
way around.  This will permit use of the advective form of thickness diffusivity
while keeping the density-triad implementation of isoneutral diffusivity.
The complementary case (skew-flux GM with an alternate isoneutral diffusion scheme)
is not likely to be run. Future commits to allow the two diffusivities
to vary independently will take care of that case.

2. The proportionality coefficient relating the thickness and isoneutral
diffusivities is passed as an argument to GMKDIF.

3. Corrected a sign error in FXZ, FYZ which did not affect the case when
the two diffusivities were taken to be equal.