Several fixes to enable unit tests to work with gcc.
1) Default ('*') formatting produces spurious blanks that the tests do not
expect.  Switched to '(a)' to avoid this issue.
2) Ensured initialization of a few components in data structures.
3) Bug fix for situation where Intel really should have caught.   Comparing
two different data types with "==", with a scalar on the left and a 1D array
on right.   "==" has been overloaded to deal with scalar-scalar and
array-array cases.   No idea what ifort was doing before this fix.
(Apparently the "right" thing, but still scary.)