Unfortunately the tracer infrastructure change yesterday tripped a compiler bug in gfortran.   Previously, TRACER_COM was "re-exporting" entities from OldTracer_mod, but the compiler now chokes on this with internal compiler errors, and I could not find a simple workaround.
Instead, I've gone through the model and replaced many "USE TRACER_COM" with
"USE OldTracer_mod".   In some cases, I used this opportunity to also fix USE
statements that were split across multiple CPP directives, so it was a net
win, but in other cases this was merely tedious.

Hopefully most of this will disappear again with the next push.

0 diffs.