This is an attempt to add an outline for new model documentation. Two documents are being added:
ModelDescription
UserGuide

Most sections in ModelDescription are present but probaly need to be
updated. Sections with new functionality need to be added.

A lot of sections in UserGuide are empty and need to be written.
(Some of them are present in old documentation, just didn't have
time to convert them yet - will do it soon). Some sections, like
"Setting up a particular modelE simulation" need to be extended
to include new functionality.