Collection of changes for more dynamic implementation of tracers within the model.
============================================
*** IMPORTANT NOTE for tracer developers ***
============================================
I tested thes changes thoroughly against a large number of
rundecks, but there remains a reasonable chance that any tracer
combination not tested will either fail to compile or have a more
subtle error due to implicit assumptions about the order of
initialization.   Fixes are usually easy once detected.
============================================

This is a collection of changes that are aimed at allowing a more
dynamic implementation of tracers within the model.  In particular, it
represents a migratino of many tracer spceficitions to _before_ allocation
of arrays dimensioned with NTM.

1) Split initTracerMetadata() into two procedures.  One is called before
   allocations, while the other is called at the original point.  (2nd name
   "lateInitTraceMetadata()" should be improved in a later patch.)

2) Minor corrections to enable change (1):
   - moved itime_tr0 initialization to just before 1st use.
   - relocated iso_index into OldTracer_mod (as with all NTM sized arrays)
   - Pulled out calls to LOGINTP into separate procedure and placed
     that call later in the initialization process.

3) Uninitialized (and unused array) TRSNOWBV0, now initialized.  (Will
   be removed in a later patch.)