Enabling E4TampF40 rundeck.
This commit is for a variety of minor fixes to restore the ability for
AMP tracers to work for the development branch of the model, as
determined by the E4TampF40 rundeck.

1) Adding/moving USE statements to make entities accessible in (presumably) new locations.

2) Added function aerosolKind() that pulls out appropriate substring
from tracer name and replaced the logic which uses the substring with
a select case.  (Necessary because trname() is now a function and
expressions like trname(n)(6:8) no longer are permitted.)

3) Added empty module to work around intel internal compiler error
that we thought was eliminated by the 13.0 release, but apparently
not.