In this commit we have extracted tracer metadata setup routines from initTracerMetada.f and created modules that set up tracer metadata for each of the following tracer subgroups: TOMAS, AMP, Shindell, Koch, Lerner, Minerals and Quarzhem.
Additionally a sharedTracersMetadata module has been created to allow for
sharing of common (duplicate) setup routines among tracer subgroups.

The following rundecks have been modified and tested for reproducibility
with the master branch baseline:

        E4TcadiF40 : Shindell, Koch
        E4TcampiF40 : Shindell, AMP
        E4TctomasF40 : Shindell, TOMAS
        E4TampF40 : AMP

Currently there are no working rundecks for Lerner, Minerals and Quarzhem
but we hope to be able to test those in the future.