These modifications should allow the model to work with netcdf and openmpi packages installed by yum (on Linux).
For openmpi one doesn't need to specify MPIDIR in this case.
Instead, one should load a corresponding module (as we do
on discover), i.e. one should execute something like

  module load openmpi-x86_64

This was tesed on Fedora 17. On CentOS it is likely that
yum-installed packages will not work with latest gcc compiler,
since CentOS uses very old compiler to cpompile its packages.