Major extension/simplification of tracer infrastructure.
This commit represents the completion of several important
modules/classes for supporting collections of tracers in the model.
None of this new logic is yet exercised in the parameterizations, but
the interfaces are mature and stable and should now be deployed.

Far more extensive documentation and training materials are being
developed, but the primary new features/capabilities are:

* Associative Arrays

A new CPP template has been created that can be used to create an
associative array for storing collections of user-defined types.
Think hash in Perl or dict in python.  The instantiated modules are
strongly typed so that one can ensure that only certain types of
objects are contained.  (See below for specific examples.)  This
package is of very general value, and other opportunities for
exploiting it will no doubt arise.

Note that F2003 also allows a completely untyped associative array
which would avoid the need for a CPP template, but I've found that
less useful for modelE, and gfortran still has some limitations with
those features.

Typical usage looks like the following:

!--------------
  type (FooAssociativeArray) :: myArray
  type (Foo) :: myObject
  class (Foo), pointer :: anObject

  myArray = newFooAssociativeArray()

  call myArray%insert('objectName', myObject) ...
  ...
  myObject => myArray%find('objectName')
  ...
!--------------

* Attributes

This is a new mechanism which replaces the previous (also relatively
new) mechanism GenericType_mod, by making heavier use of F2003 OO
capabilities.  Each intrinsic type is now wrapped in a derived type
which extends the base AbstractAttribute type.  Assignment(=) is
overloaded for convenience.

The m4 preprocessor is used for overloading the various types:
  - IntegerAttribute, Integer1DAttribute
  - RealDPAttribute, RealDP1DAttribute
  - LogicalAttribute, Logical1DAttribute
  - StringAttribute, String1DAttribute

The module Attributes_mod bundles these "sub" modules together for convenience.

An additional module which is for convenience when dealing with
tracer bundles is AttributeReference.F90 which allows arrays of
pointers of attributes (as opposed to a pointer to an array of
attributes).  There is also a VectorAttribute derived type that
extends AbstractAttribute for convenience of storing such arrays inside
an appropriate associative array  (which only knows about scalars).

* AttributeDictionary.F90 extends the AttributeAssociativeArray class
(derived from template), to overload the insert() method for the
various intrinsic types.  Only necessary because the intrinsics
cannot be treated as objects directly.

* Tracer.F90 has been updated to EXTEND the AttributeDictionary type
which eliminates a large fraction of the previous logic.

* As with Tracer.F90, TracerBundle.F90 EXTENDs the
TracerAssociativeArray class (derived from template), but also
includes logic for managing default attributes, mandatory
attributes, bundle subsets, and accessors for directly getting
attributes from tracers.  (Otherwise is a 2-step process.)  For
performance purposes, additional logic and datatypes have been
introduced to obtain 1D array of attributes corresponding to 1
attribute per tracer.  These will be the ultimate replacement of the
original tracer arrays such as:  tr_mm, mass2vol, etc.

Other changes include:

1) Splitting contents of AttributeDictionary.m4
Attribute classes for each intrinsic type are now in Attributes.m4F90
along with a module that re-exports all of them.  This leaves
AttributeDictionary with just the contents that the name implies.

2) Added Hash template to augment template for Associative Arrays.
AttributeDictionary now extends hash, which implies that Tracer objects are now Hash's.
TracerBundle now extends Hash rather than AssociativeArray
Eliminated/corrected some legacy interfaces based upon absolute index
of a tracer.  No longer safe to use iterator with counter for a Hash.

3) Corrected problem with hasNext()/next() for iterators of empty Hash's.

4) Added unit tests for hash to expose/fix this problem.  Had skipped TDD
when duplicating Hash from Associative arrays.

5) Transitioning to allocatable components for HashMaps.
This should be safer for memory management.
Constructors no longer return pointers.

6) Significant progress to use ALLOCATABLE components in tracer
infrastructure instead of POINTER components.  Although safer for
memory management, ALLOCATABLE appears to require care for working
around compiler issues.

7) Miscellaneous refactoring