Converted files OIC, TOPO_OC, and KBASIN to netcdf format.  This commit introduces roundoff differences due to the elimination of the use of 4-byte real arrays in certain cold-start computations.
The KBASIN file now contains zero/one masks for each of the
named ocean basins (Pacific, Atlantic, Indian).  This file
was previously ASCII with a letter code.  This change allows
offline programs/scripts to more easily calculate various
basin averages identically to the main model.

This commit proved to be an opportune time to eliminate, in preparation
for future changes, a number of global-sized arrays that were never
converted to local size during the initial domain decomposition work.