Now computing the polar filtering coefficients during initialization rather than reading them from an input file.  This eliminates a longstanding potential source of inconsistencies and frustration when changing the bathymetry or resolution.  On one processor, the total time to compute the coefficients is roughly 1 second for 288x180x32 resolution (after various optimizations to the original coding).
The new version of the filtering routine does not yield exactly the
same coefficients as the current input files, for three reasons:
(1) The offline calculation is not calculating DXP the same as the OGCM.
    I did not try to reproduce the offline DXP in the model.
(2) The new version applies the filter at latitudes for which DXP<DYP,
    while the old version used 40 degrees latitude as the cutoff.
(3) Roundoff differences (e.g. no more 4-byte to 8-byte conversions).