Introduced a mechanism to set some relevant parameters in the constants module to other values if simulating an environment other than modern Earth.
The global mean surface pressure can also be set via this
mechanism even though it resides in the resolution module
rather than the constants module (I may change this soon).

CPP-defining PLANET_PARAMS to a particular string (i.e. planet name)
in the rundeck will cause the collection of parameters for that planet
to be used.  These collections of parameters are defined in new
file PlanetParams_mod.F90.

Leaving all these parameters as compile-time constants
minimized disruption in the rest of the code.  While the
exercise of making the necessary changes to accommodate the
loss of compile-time information did not indicate that
the disruption would be large, I decided to postpone this
until some threshold is reached in the fraction of the user base
needing to set various parameters independently through the
rundeck.